pretty sure i'm just about done with zod validation. not done with functions themselves though ofc
