--- firebase-project/functions/src/models/location.model.ts
@@ -1,11 +1,0 @@
-import { z } from "zod";
-
-export const locationSchema = z.object({
-  latitude: z.number(),
-  longitude: z.number(),
-  geohash: z.string(),
-  city: z.string().optional(),
-  relativeDistance: z.number().optional(),
-});
-
-export type Location = z.infer<typeof locationSchema>;
