--- firebase-project/functions/src/models/utils.model.ts
@@ -1,4 +1,0 @@
-import { z } from "zod";
-
-export const documentIdSchema = z.string().length(28);
-export type DocumentIdType = z.infer<typeof documentIdSchema>;
